lib/atoi/, *: Split files
