+++ lib/Makefile.am
@@ -43,16 +43,34 @@ libshadow_la_SOURCES = \
 	alloc/x/xmalloc.h \
 	alloc/x/xrealloc.c \
 	alloc/x/xrealloc.h \
-	atoi/a2i.c \
-	atoi/a2i.h \
+	atoi/a2i/a2i.c \
+	atoi/a2i/a2i.h \
+	atoi/a2i/a2s.c \
+	atoi/a2i/a2s.h \
+	atoi/a2i/a2s_c.c \
+	atoi/a2i/a2s_c.h \
+	atoi/a2i/a2s_nc.c \
+	atoi/a2i/a2s_nc.h \
+	atoi/a2i/a2u.c \
+	atoi/a2i/a2u.h \
+	atoi/a2i/a2u_c.c \
+	atoi/a2i/a2u_c.h \
+	atoi/a2i/a2u_nc.c \
+	atoi/a2i/a2u_nc.h \
 	atoi/getnum.c \
 	atoi/getnum.h \
-	atoi/str2i.c \
-	atoi/str2i.h \
-	atoi/strtoi.c \
-	atoi/strtoi.h \
-	atoi/strtou_noneg.c \
-	atoi/strtou_noneg.h \
+	atoi/str2i/str2i.c \
+	atoi/str2i/str2i.h \
+	atoi/str2i/str2s.c \
+	atoi/str2i/str2s.h \
+	atoi/str2i/str2u.c \
+	atoi/str2i/str2u.h \
+	atoi/strtoi/strtoi.c \
+	atoi/strtoi/strtoi.h \
+	atoi/strtoi/strtou.c \
+	atoi/strtoi/strtou.h \
+	atoi/strtoi/strtou_noneg.c \
+	atoi/strtoi/strtou_noneg.h \
 	attr.h \
 	audit_help.c \
 	basename.c \
