+++ lib/atoi/getnum.h
@@ -13,7 +13,8 @@
 #include <stddef.h>
 #include <sys/types.h>
 
-#include "atoi/a2i.h"
+#include "atoi/a2i/a2i.h"
+#include "atoi/a2i/a2s.h"
 #include "attr.h"
 #include "typetraits.h"
