+++ lib/atoi/str2i/str2u.c
@@ -5,13 +5,9 @@
 
 #include <config.h>
 
-#include "atoi/str2i.h"
+#include "atoi/str2i/str2u.h"
 
 
-extern inline int str2sh(short *restrict n, const char *restrict s);
-extern inline int str2si(int *restrict n, const char *restrict s);
-extern inline int str2sl(long *restrict n, const char *restrict s);
-extern inline int str2sll(long long *restrict n, const char *restrict s);
 extern inline int str2uh(unsigned short *restrict n, const char *restrict s);
 extern inline int str2ui(unsigned int *restrict n, const char *restrict s);
 extern inline int str2ul(unsigned long *restrict n, const char *restrict s);
