+++ lib/atoi/strtoi/strtou_noneg.c
@@ -1,10 +1,10 @@
-// SPDX-FileCopyrightText: 2023, Alejandro Colomar <alx@kernel.org>
+// SPDX-FileCopyrightText: 2023-2024, Alejandro Colomar <alx@kernel.org>
 // SPDX-License-Identifier: BSD-3-Clause
 
 
 #include <config.h>
 
-#include "atoi/strtou_noneg.h"
+#include "atoi/strtoi/strtou_noneg.h"
 
 #include <stdint.h>
