+++ lib/getdef.c
@@ -22,8 +22,9 @@
 #include <libeconf.h>
 #endif
 
-#include "atoi/a2i.h"
-#include "atoi/str2i.h"
+#include "atoi/a2i/a2s.h"
+#include "atoi/a2i/a2u.h"
+#include "atoi/str2i/str2u.h"
 #include "defines.h"
 #include "getdef.h"
 #include "prototypes.h"
