+++ lib/getrange.c
@@ -10,7 +10,7 @@
 #include <ctype.h>
 #include <stdlib.h>
 
-#include "atoi/a2i.h"
+#include "atoi/a2i/a2u.h"
 #include "defines.h"
 #include "prototypes.h"
