+++ lib/limits.c
@@ -30,8 +30,11 @@
 #include "shadowlog.h"
 #include <sys/resource.h>
 
-#include "atoi/a2i.h"
-#include "atoi/str2i.h"
+#include "atoi/a2i/a2i.h"
+#include "atoi/a2i/a2s.h"
+#include "atoi/str2i/str2i.h"
+#include "atoi/str2i/str2s.h"
+#include "atoi/str2i/str2u.h"
 #include "memzero.h"
 #include "typetraits.h"
