+++ lib/shadow.c
@@ -18,8 +18,8 @@
 #include <string.h>
 #include <sys/types.h>
 
-#include "atoi/a2i.h"
-#include "atoi/str2i.h"
+#include "atoi/a2i/a2s.h"
+#include "atoi/str2i/str2u.h"
 #include "defines.h"
 #include "prototypes.h"
 #include "string/strtok/stpsep.h"
