+++ src/new_subid_range.c
@@ -3,7 +3,7 @@
 #include <stdio.h>
 #include <unistd.h>
 
-#include "atoi/str2i.h"
+#include "atoi/str2i/str2u.h"
 #include "subid.h"
 #include "stdlib.h"
 #include "prototypes.h"
