+++ src/newusers.c
@@ -32,7 +32,7 @@
 
 #include "alloc/reallocf.h"
 #include "atoi/getnum.h"
-#include "atoi/str2i.h"
+#include "atoi/str2i/str2s.h"
 #ifdef ACCT_TOOLS_SETUID
 #ifdef USE_PAM
 #include "pam_defs.h"
