+++ src/useradd.c
@@ -37,7 +37,7 @@
 #include <unistd.h>
 
 #include "alloc/x/xmalloc.h"
-#include "atoi/a2i.h"
+#include "atoi/a2i/a2s.h"
 #include "atoi/getnum.h"
 #include "chkname.h"
 #include "defines.h"
