+++ src/usermod.c
@@ -34,7 +34,8 @@
 
 #include "alloc/malloc.h"
 #include "alloc/x/xmalloc.h"
-#include "atoi/a2i.h"
+#include "atoi/a2i/a2i.h"
+#include "atoi/a2i/a2s.h"
 #include "atoi/getnum.h"
 #include "chkname.h"
 #include "defines.h"
