+++ tests/unit/Makefile.am
@@ -35,7 +35,8 @@ test_adds_LDADD = \
     $(NULL)
 
 test_atoi_strtoi_SOURCES = \
-    ../../lib/atoi/strtoi.c \
+    ../../lib/atoi/strtoi/strtoi.c \
+    ../../lib/atoi/strtoi/strtou_noneg.c \
     test_atoi_strtoi.c \
     $(NULL)
 test_atoi_strtoi_CFLAGS = \
