+++ tests/unit/test_atoi_strtoi.c
@@ -14,8 +14,8 @@
 #include <stdint.h>  // Required by <cmocka.h>
 #include <cmocka.h>
 
-#include "atoi/strtoi.h"
-#include "atoi/strtou_noneg.h"
+#include "atoi/strtoi/strtoi.h"
+#include "atoi/strtoi/strtou_noneg.h"
 
 
 static void test_strtoi(void **state);
